backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-cloud-speech==2.29.0
google-cloud-texttospeech==2.20.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.0
